refactor(#559): refactor RemoteWITService as per new design of service
