+++ application/service/services.go
@@ -17,9 +17,9 @@ import (
 
 /*
 Steps for adding a new Service:
-1. Add a new service interface to application/service/service.go
+1. Add a new service interface to application/service/services.go
 2. Create an implementation of the service interface
-3. Add a new method to service.Service interface in application/service/service.go for accessing the service interface
+3. Add a new method to service.Services interface in application/service/services.go for accessing the service interface
    defined in step 1
 4. Add a new method to application/service/factory/service_factory.go which implements the service access method
    from step #3 and uses the service constructor from step 2
@@ -79,6 +79,12 @@ type NotificationService interface {
 	SendMessagesAsync(ctx context.Context, messages []notification.Message) error
 }
 
+type WITService interface {
+	UpdateWITUser(ctx context.Context, updatePayload *app.UpdateUsersPayload, identityID string) error
+	CreateWITUser(ctx context.Context, identity *account.Identity, identityID string) error
+	GetSpaceNameAndOwnedBy(ctx context.Context, spaceID string) (name, ownedBy string, e error)
+}
+
 //Services creates instances of service layer objects
 type Services interface {
 	InvitationService() InvitationService
@@ -90,4 +96,5 @@ type Services interface {
 	SpaceService() SpaceService
 	UserService() UserService
 	NotificationService() NotificationService
+	WITService() WITService
 }
