+++ controller/users.go
@@ -22,13 +22,13 @@ import (
 	"github.com/fabric8-services/fabric8-auth/login/link"
 	"github.com/fabric8-services/fabric8-auth/rest"
 	"github.com/fabric8-services/fabric8-auth/token"
-	"github.com/fabric8-services/fabric8-auth/wit"
-
 	"github.com/goadesign/goa"
 	"github.com/goadesign/goa/middleware/security/jwt"
 	"github.com/jinzhu/gorm"
 	errs "github.com/pkg/errors"
 	"github.com/satori/go.uuid"
+	svc "github.com/fabric8-services/fabric8-auth/application/service"
+
 )
 
 // UsersController implements the users resource.
@@ -37,7 +37,7 @@ type UsersController struct {
 	app                      application.Application
 	config                   UsersControllerConfiguration
 	userProfileService       login.UserProfileService
-	RemoteWITService         wit.RemoteWITService
+	WITService               svc.WITService
 	EmailVerificationService service.EmailVerificationService
 	keycloakLinkService      link.KeycloakIDPService
 }
@@ -65,7 +65,7 @@ func NewUsersController(service *goa.Service, app application.Application, confi
 		app:                 app,
 		config:              config,
 		userProfileService:  userProfileService,
-		RemoteWITService:    &wit.RemoteWITServiceCaller{},
+		WITService:          app.WITService(),
 		keycloakLinkService: linkService,
 	}
 }
@@ -177,17 +177,7 @@ func (c *UsersController) Create(ctx *app.CreateUsersContext) error {
 	}
 
 	// finally, if all works, we create a user in WIT too.
-	witURL, err := c.config.GetWITURL()
-	if err != nil {
-		log.Error(ctx, map[string]interface{}{
-			"err":              err,
-			"keycloak_user_id": *keycloakUserID,
-		}, "failed to create user in WIT")
-
-		return jsonapi.JSONErrorResponse(ctx, errors.NewInternalError(ctx, err))
-	}
-
-	err = c.RemoteWITService.CreateWITUser(ctx.Context, identity, witURL, identityID.String())
+	err = c.WITService.CreateWITUser(ctx.Context, identity, identityID.String())
 	if err != nil {
 		log.Error(ctx, map[string]interface{}{
 			"err":              err,
@@ -700,8 +690,8 @@ func (c *UsersController) Update(ctx *app.UpdateUsersContext) error {
 			switch err.(type) {
 			default:
 				return ctx.BadRequest(jerrors)
-			// case errors.BadParameterError:
-			// 	return ctx.Conflict(jerrors)
+				// case errors.BadParameterError:
+				// 	return ctx.Conflict(jerrors)
 			case errors.UnauthorizedError:
 				return ctx.Unauthorized(jerrors)
 			}
@@ -737,7 +727,7 @@ func (c *UsersController) updateFeatureLevel(ctx context.Context, user *accountr
 		} else {
 			// if the level is 'internal', we need to check against the email address to verify that the user is a Red Hat employee
 			if *updatedFeatureLevel == "internal" &&
-				// do not allow if email is not verified or if email belongs to another domain
+			// do not allow if email is not verified or if email belongs to another domain
 				(!user.EmailVerified || !strings.HasSuffix(user.Email, c.config.GetInternalUsersEmailAddressSuffix())) {
 				log.Error(ctx, map[string]interface{}{"user_id": user.ID, "user_email": user.Email}, "user is not an employee")
 				return errors.NewForbiddenError("User is not allowed to opt-in for the 'internal' level of features.")
@@ -759,17 +749,14 @@ func (c *UsersController) updateWITUser(ctx *app.UpdateUsersContext, identityID 
 				FullName:              ctx.Payload.Data.Attributes.FullName,
 				ImageURL:              ctx.Payload.Data.Attributes.ImageURL,
 				RegistrationCompleted: ctx.Payload.Data.Attributes.RegistrationCompleted,
-				URL:      ctx.Payload.Data.Attributes.URL,
-				Username: ctx.Payload.Data.Attributes.Username,
+				URL:                   ctx.Payload.Data.Attributes.URL,
+				Username:              ctx.Payload.Data.Attributes.Username,
 			},
 			Type: ctx.Payload.Data.Type,
 		},
 	}
-	witURL, err := c.config.GetWITURL()
-	if err != nil {
-		return err
-	}
-	return c.RemoteWITService.UpdateWITUser(ctx, updateUserPayload, witURL, identityID)
+
+	return c.WITService.UpdateWITUser(ctx, updateUserPayload, identityID)
 }
 
 func isEmailValid(email string) bool {
